requirements.txt
accelerate==1.12.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_43o9lm92p9/croot/certifi_1750863864870/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
emoji==2.14.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools==4.61.0
fsspec==2025.5.1
ftfy==6.3.1
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
huggingface-hub==0.36.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
kiwisolver==1.4.9
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.8
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==3.2.5
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2024.11.6
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
safetensors==0.7.0
scipy==1.16.3
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==72.1.0
six==1.17.0
sympy==1.13.1
timm==1.0.22
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.3
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
ultralytics==8.3.235
ultralytics-thop==2.0.18
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
wcwidth==0.2.14
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
